claudeagentsdk-main/claudeagentsdk-main/requirements.txt
claude-agent-sdk>=0.1.0
mem0ai>=0.1.0
supabase>=2.7.0
python-dotenv>=1.0.0
